+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `OutTextForUnit` API
+- `AddStaticObject` API (for standard static objects)
+- `AddLinkedStatic` API (for statics linked to units such as ships)
 - `MarkupToAll` API
 - `MarkupToCoalition` API
 - `GetTheatre` API